refactor: structure basic components outside edit pictures section
